+++ internal/prototest/examplepb/example.pb.yarpc.go
@@ -199,12 +199,18 @@ func NewFxKeyValueYARPCProcedures() interface{} {
 				Server:      params.Server,
 				AnyResolver: params.AnyResolver,
 			}),
-			ReflectionMeta: reflection.ServerMeta{
-				ServiceName:     "uber.yarpc.internal.examples.protobuf.example.KeyValue",
-				FileDescriptors: yarpcFileDescriptorClosurebab15b635bbc13f7,
-			},
+			ReflectionMeta: KeyValueReflectionMeta,
 		}
 	}
+}
+
+// KeyValueReflectionMeta is the reflection server metadata
+// required for using the gRPC reflection protocol with YARPC.
+//
+// See https://github.com/grpc/grpc/blob/master/doc/server-reflection.md.
+var KeyValueReflectionMeta = reflection.ServerMeta{
+	ServiceName:     "uber.yarpc.internal.examples.protobuf.example.KeyValue",
+	FileDescriptors: yarpcFileDescriptorClosurebab15b635bbc13f7,
 }
 
 type _KeyValueYARPCCaller struct {
@@ -502,12 +508,18 @@ func NewFxFooYARPCProcedures() interface{} {
 				Server:      params.Server,
 				AnyResolver: params.AnyResolver,
 			}),
-			ReflectionMeta: reflection.ServerMeta{
-				ServiceName:     "uber.yarpc.internal.examples.protobuf.example.Foo",
-				FileDescriptors: yarpcFileDescriptorClosurebab15b635bbc13f7,
-			},
+			ReflectionMeta: FooReflectionMeta,
 		}
 	}
+}
+
+// FooReflectionMeta is the reflection server metadata
+// required for using the gRPC reflection protocol with YARPC.
+//
+// See https://github.com/grpc/grpc/blob/master/doc/server-reflection.md.
+var FooReflectionMeta = reflection.ServerMeta{
+	ServiceName:     "uber.yarpc.internal.examples.protobuf.example.Foo",
+	FileDescriptors: yarpcFileDescriptorClosurebab15b635bbc13f7,
 }
 
 type _FooYARPCCaller struct {
@@ -885,12 +897,18 @@ func NewFxTestMessageNameParityYARPCProcedures() interface{} {
 				Server:      params.Server,
 				AnyResolver: params.AnyResolver,
 			}),
-			ReflectionMeta: reflection.ServerMeta{
-				ServiceName:     "uber.yarpc.internal.examples.protobuf.example.TestMessageNameParity",
-				FileDescriptors: yarpcFileDescriptorClosurebab15b635bbc13f7,
-			},
+			ReflectionMeta: TestMessageNameParityReflectionMeta,
 		}
 	}
+}
+
+// TestMessageNameParityReflectionMeta is the reflection server metadata
+// required for using the gRPC reflection protocol with YARPC.
+//
+// See https://github.com/grpc/grpc/blob/master/doc/server-reflection.md.
+var TestMessageNameParityReflectionMeta = reflection.ServerMeta{
+	ServiceName:     "uber.yarpc.internal.examples.protobuf.example.TestMessageNameParity",
+	FileDescriptors: yarpcFileDescriptorClosurebab15b635bbc13f7,
 }
 
 type _TestMessageNameParityYARPCCaller struct {